backend/requirements.txt
aniso8601==9.0.1
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
cryptography==41.0.3
defusedxml==0.7.1
Deprecated==1.2.14
dj-database-url==2.1.0
dj-email-url==1.0.6
dj-rest-auth==4.0.1
Django==4.2.4
django-allauth==0.54.0
django-cache-url==3.4.4
django-debug-toolbar==3.4.0
django-memcached==0.1.2
django-redis==5.3.0
django-rest-auth==0.9.5
django-taggit==4.0.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-spectacular==0.26.4
environs==9.5.0
graphene==3.3
graphene-django==3.1.5
graphql-core==3.2.3
graphql-relay==3.2.0
idna==3.4
inflection==0.5.1
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
marshmallow==3.20.1
oauthlib==3.2.2
packaging==23.1
Pillow==10.0.0
promise==2.3
psycopg2-binary==2.9.3
pycparser==2.21
PyJWT==2.8.0
pymemcache==4.0.0
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2023.3
PyYAML==6.0.1
redis==4.3.4
referencing==0.30.2
requests==2.31.0
requests-oauthlib==1.3.1
rpds-py==0.10.0
Rx==1.6.3
singledispatch==4.1.0
six==1.16.0
sqlparse==0.4.4
text-unidecode==1.3
typing_extensions==4.7.1
tzdata==2023.3
Unidecode==1.1.1
uritemplate==4.1.1
urllib3==2.0.4
wrapt==1.15.0
django-elasticsearch-dsl==7.3 
django-elasticsearch-dsl-drf==0.22.5 